+++ src_Set-OutlookSignatures/sample code/Create-DesktopIcon.ps1
@@ -10,6 +10,12 @@ Would you like support? ExplicIT Consulting (https://explicitconsulting.at) offe
 
 [CmdletBinding()] param ()
 
+if ($psISE) {
+    Write-Host 'PowerShell ISE detected. Use PowerShell in console or terminal instead.' -ForegroundColor Red
+    Write-Host 'Required features are not available in ISE. Exit.' -ForegroundColor Red
+    exit 1
+}
+
 $OutputEncoding = [Console]::InputEncoding = [Console]::OutputEncoding = New-Object System.Text.UTF8Encoding
 
 Set-Location $PSScriptRoot