+++ src_Set-OutlookSignatures/sample code/SimulationModeHelper.ps1
@@ -25,6 +25,18 @@ $params = [ordered]@{
 	GraphOnly                     = $false
 }
 
+
+#
+# Do not change anything from here on
+#
+
+
+if ($psISE) {
+	Write-Host '  PowerShell ISE detected. Use PowerShell in console or terminal instead.' -ForegroundColor Red
+	Write-Host '  Required features are not available in ISE. Exit.' -ForegroundColor Red
+	exit 1
+}
+
 $OutputEncoding = [Console]::InputEncoding = [Console]::OutputEncoding = New-Object System.Text.UTF8Encoding
 
 Set-Location $PSScriptRoot